+++ source/ir/Zambezi/buffer/NewBufferMaps.cpp
@@ -33,7 +33,7 @@ void NewBufferMaps::save(std::ostream& ostr) const
         ostr.write((const char*)&score_[i][0], sizeof(uint32_t) * size);
     }
 
-    ostr.write((const char*)&tailPointer_[0], sizeof(uint32_t) * capacity_);
+    ostr.write((const char*)&tailPointer_[0], sizeof(size_t) * capacity_);
 }
 
 void NewBufferMaps::load(std::istream& istr)
